Move `SheetPos` and `SheetRect` to `position.rs`
