+++ quadratic-core/src/controller/compute.rs
@@ -6,13 +6,11 @@ use super::{
 use crate::{
     grid::{
         js_types::{CellForArray, JsRenderCellUpdate, JsRenderCellUpdateEnum},
-        CodeCellLanguage, CodeCellRunOutput, CodeCellRunResult, CodeCellValue, SheetId,
+        CodeCellLanguage, CodeCellRunOutput, CodeCellRunResult, CodeCellValue,
     },
     wasm_bindings::{js::runPython, JsComputeResult},
-    Pos, Value,
+    Pos, SheetPos, SheetRect, Value,
 };
-use serde::{Deserialize, Serialize};
-use std::{fmt, ops::Range};
 
 impl GridController {
     /// Given `cell` and `dependencies` adds a new node to the graph.
@@ -219,78 +217,10 @@ impl GridController {
     }
 }
 
-/// Used for referencing a range during computation.
-#[derive(Serialize, Deserialize, Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub struct SheetRect {
-    /// Upper-left corner.
-    pub min: Pos,
-    /// Lower-right corner.
-    pub max: Pos,
-    /// The sheet that this region is on.
-    pub sheet_id: SheetId,
-}
-
-impl SheetRect {
-    /// Constructs a new rectangle containing only a single cell.
-    pub fn single_pos(pos: SheetPos) -> SheetRect {
-        SheetRect {
-            sheet_id: pos.sheet_id,
-            min: Pos { x: pos.x, y: pos.y },
-            max: Pos { x: pos.x, y: pos.y },
-        }
-    }
-    /// Returns whether a position is contained within the rectangle.
-    pub fn contains(self, pos: SheetPos) -> bool {
-        self.sheet_id == pos.sheet_id
-            && self.x_range().contains(&pos.x)
-            && self.y_range().contains(&pos.y)
-    }
-    /// Returns whether a rectangle intersects with the rectangle.
-    pub fn intersects(self, other: SheetRect) -> bool {
-        // https://en.wikipedia.org/wiki/Hyperplane_separation_theorem#:~:text=the%20following%20form%3A-,Separating%20axis%20theorem,-%E2%80%94%C2%A0Two%20closed
-        self.sheet_id == other.sheet_id
-            && !(other.max.x < self.min.x
-                || other.min.x > self.max.x
-                || other.max.y < self.min.y
-                || other.min.y > self.max.y)
-    }
-    /// Returns the range of X values in the rectangle.
-    pub fn x_range(self) -> Range<i64> {
-        self.min.x..self.max.x + 1
-    }
-    /// Returns the range of Y values in the rectangle.
-    pub fn y_range(self) -> Range<i64> {
-        self.min.y..self.max.y + 1
-    }
-}
-impl fmt::Display for SheetRect {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "Sheet: {}, Min: {}, Max: {}",
-            self.sheet_id, self.min, self.max,
-        )
-    }
-}
-
-/// Used for referencing a pos during computation.
-#[derive(Serialize, Deserialize, Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub struct SheetPos {
-    pub x: i64,
-    pub y: i64,
-    pub sheet_id: SheetId,
-}
-impl fmt::Display for SheetPos {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{} ({}, {})", self.sheet_id, self.x, self.y)
-    }
-}
-
 #[cfg(test)]
 mod test {
     use crate::controller::{transaction_summary::TransactionSummary, GridController};
-
-    use super::{SheetPos, SheetRect};
+    use crate::{SheetPos, SheetRect};
 
     #[tokio::test]
     async fn test_graph() {
