+++ quadratic-core/src/position.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 #[cfg(feature = "js")]
 use wasm_bindgen::prelude::*;
 
-use crate::{controller::compute::SheetPos, ArraySize};
+use crate::{grid::SheetId, ArraySize};
 
 /// Cell position {x, y}.
 #[cfg_attr(test, derive(proptest_derive::Arbitrary))]
@@ -14,7 +14,6 @@ use crate::{controller::compute::SheetPos, ArraySize};
 )]
 #[cfg_attr(feature = "js", wasm_bindgen)]
 #[cfg_attr(feature = "js", derive(ts_rs::TS))]
-
 pub struct Pos {
     /// Column
     #[cfg_attr(test, proptest(strategy = "-4..=4_i64"))]
@@ -156,4 +155,70 @@ pub struct ScreenRect {
     pub y: f64,
     pub w: f64,
     pub h: f64,
+}
+
+/// Used for referencing a pos during computation.
+#[derive(Serialize, Deserialize, Debug, Copy, Clone, PartialEq, Eq, Hash)]
+pub struct SheetPos {
+    pub x: i64,
+    pub y: i64,
+    pub sheet_id: SheetId,
+}
+impl fmt::Display for SheetPos {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{} ({}, {})", self.sheet_id, self.x, self.y)
+    }
+}
+
+/// Used for referencing a range during computation.
+#[derive(Serialize, Deserialize, Debug, Copy, Clone, PartialEq, Eq, Hash)]
+pub struct SheetRect {
+    /// Upper-left corner.
+    pub min: Pos,
+    /// Lower-right corner.
+    pub max: Pos,
+    /// The sheet that this region is on.
+    pub sheet_id: SheetId,
+}
+impl SheetRect {
+    /// Constructs a new rectangle containing only a single cell.
+    pub fn single_pos(pos: SheetPos) -> SheetRect {
+        SheetRect {
+            sheet_id: pos.sheet_id,
+            min: Pos { x: pos.x, y: pos.y },
+            max: Pos { x: pos.x, y: pos.y },
+        }
+    }
+    /// Returns whether a position is contained within the rectangle.
+    pub fn contains(self, pos: SheetPos) -> bool {
+        self.sheet_id == pos.sheet_id
+            && self.x_range().contains(&pos.x)
+            && self.y_range().contains(&pos.y)
+    }
+    /// Returns whether a rectangle intersects with the rectangle.
+    pub fn intersects(self, other: SheetRect) -> bool {
+        // https://en.wikipedia.org/wiki/Hyperplane_separation_theorem#:~:text=the%20following%20form%3A-,Separating%20axis%20theorem,-%E2%80%94%C2%A0Two%20closed
+        self.sheet_id == other.sheet_id
+            && !(other.max.x < self.min.x
+                || other.min.x > self.max.x
+                || other.max.y < self.min.y
+                || other.min.y > self.max.y)
+    }
+    /// Returns the range of X values in the rectangle.
+    pub fn x_range(self) -> Range<i64> {
+        self.min.x..self.max.x + 1
+    }
+    /// Returns the range of Y values in the rectangle.
+    pub fn y_range(self) -> Range<i64> {
+        self.min.y..self.max.y + 1
+    }
+}
+impl fmt::Display for SheetRect {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "Sheet: {}, Min: {}, Max: {}",
+            self.sheet_id, self.min, self.max,
+        )
+    }
 }
